Display multiple boards in parallel
Now you can add many boards with a parallel game for them.
This is done to avoid a mistake with a single board, stored
as a global variable. The problem is that with each visit
to the site, processing of a new board was launched and the
data was mixed